Nehal 4 commit 12-12-2019

+++ app/src/main/java/com/mezyapps/bill_app/view/fragment/HomeFragment.java
@@ -238,9 +238,9 @@ public class HomeFragment extends Fragment implements SelectBillIItemInterface {
                 if (rate.equalsIgnoreCase("")) {
                     rate = "0";
                 }
-                Double qtyInt = Double.parseDouble(qty);
-                Double rateInt = Double.parseDouble(rate);
-                Double amount = qtyInt * rateInt;
+                float qtyInt = Float.parseFloat(qty);
+                float rateInt = Float.parseFloat(rate);
+                float amount = qtyInt * rateInt;
                 amt = String.valueOf(amount);
                 edt_amt.setText(amt);
             }
@@ -267,9 +267,9 @@ public class HomeFragment extends Fragment implements SelectBillIItemInterface {
                 if (rate.equalsIgnoreCase("")) {
                     rate = "0";
                 }
-                Double qtyInt = Double.parseDouble(qty);
-                Double rateInt = Double.parseDouble(rate);
-                Double amount = qtyInt * rateInt;
+                float qtyInt = Float.parseFloat(qty);
+                float rateInt = Float.parseFloat(rate);
+                float amount = qtyInt * rateInt;
                 amt = String.valueOf(amount);
                 edt_amt.setText(amt);
